Committing completed stuff
A large portion of work on the fluid cleaner and fixer is complete; most
of what still needs to be done on those is cleaning up the Fluid library
and eventually implementing the various power APIs.  Most rendering work
on the block placer item is done, but actual functionality is pending a
refactor of the rendering code to make both rendering and block placing
cleaner.