Achievement work
There is now a more or less working achievement for discovering the
plaid plains biome (or a derivative biome, if there ever is one).  I
still need to figure out localization or something for the biome
information though.  There is also an unobtainable second achievement
which is necessary to make the achievements page work.